Thunks/gen: Move diagnostic marshaling helper to a dedicated header
